Fix: track the Idol of Moloch's imbuement properly.
In the initial implementation of the Infidel role, the Idol of Moloch
being imbued was tracked by giving it an enchantment. But in doing so,
it was subject to cancellation - if that occurred after the Idol was
imbued, the 'enchantment' would be lost and would make the game
unwinnable. The quick fix made earlier was to just not allow the Idol to
be cancelled, ever. At first I thought I'd have to make a u. bitfield to
track the Idol's imbuement, but luckily it already exists in the form of
u.uachive.amulet. If playing as an Infidel, u.uachive.amulet means the
Idol was imbued. So just swap out the enchantment bit for this bitfield.
I also reverted commit a60d231 in this commit, as making the Idol
cancel-proof is no longer necessary.

Also in this commit - if playing as an Infidel, the Idol of Moloch
cannot be put into a container, regardless of whether it's been imbued
or not.

Also includes Fix: 'potentially uninitialized local pointer' from last commit