Fix: describing barding.
When first introducing barding to the game, I had set it so the 'a'
prefix was dropped from its description. It looked wrong seeing 'a
barding'. It was also weird seeing 'You see here barding', so the
prefix 'some' has been added to its description.

I'm not 100% sure if this is correct, and looking online for a solution
hasn't been fruitful. I'm not sure whether it's 'some' or 'a set of' or
if some other prefix should be used. Right now, this is better than
nothing.