Fix: Candles could skip aging when lit and player was fast (reported by mobileuser).
This was a bit confusing, since the candle would be described as partly lit while lit - but when snuffed the candle would revert to it's base description. To fix this, we do a check when candles are snuffed to see if they are still unaged and increment the age if they still have their full age.