Fix: Disallow cloning of wish generators or markers
By cloning wands of wishing or magic lamps you can generate an
infinite amount of wishes. In SpliceHack, wands of wishing are
handled the same as in xNetHack" so you can't recharge them.

By cloning magic markers you could generate an infinite amount
of marker chargers given that you had an infinite amount of blank
scrolls.
Cloning a wand of wishing will create a wand of wonder and cloning
a magic lamp will create a magic candle.
In the case of magic markers a steel athame will be created with enchantment
between 0 and marker charges/10.