Fix robe of protection giving MC2 and no effect from dragon scales
The PROTECTION flag was dropped and the can is set to 1 to give MC1.
The bug happened in wear.c:134 when the result of the bitmask operation
below set the u.uprops[ANTIMAGIC].extrinsic to 0.
u.uprops[p].extrinsic & ~wp->w_mask;
This also happened for other colors of dragon scales.
This was reported by VaderFLAG.