Allow Und-Tor to get skilled in trident
Und-Tor was given a trident as a replacement for their starting spear
and weapon.c:skill_init set all weapons that a character is starting
with to P_BASIC while the max skill for tridents on Und is restricted.
This caused an impossible and program in disorder due to
P_MAX_SKILL(P_TRIDENT) < P_SKILL(P_TRIDENT).

Reported by Disperse.