Fix: Cure withering when polymorphing into something nonliving (EvilHack)
Being non-living prevented withering attacks from hurting you, but didn't cure existing withering.

This fix was a bit complicated because withering is an actual property that could in theory be extrinsic. I think the neatest way to futureproof for that is to use uprops[WITHERING].blocked and set it FROMFORM in set_uasmon. With BWithering defined, it might as well be used in mhitu for consistency.

Currently is_vampshifter doesn't work properly for youmonst (#monstering from vampire to second form just makes you a normal raven/&c.), so I removed that bit for now.