Allow monks with intrinsic water walking to enter into and stay in pools/moats. (EvilHack)
Monks that gained intrinsic water walking once they reached experience
level 23 discovered an odd problem. They could no longer enter a pool of
water or moat. And on the off-chance they were underwater and then made
it to level 23, they'd float up and out of the water, never to go back
in. For tortle monks, this would be a problem.

In this commit, if you have *intrinsic* water walking, you can choose to
enter a pool or moat by using the > key when standing on top of the
water tile. And while underwater, you won't float back up unless you
choose to do so.