Add Grandmaster's Robe artifact
Fixes #346.
The robe can be named from a robe of power if the character is level 30
and has grand mastery in martial arts.
The robe itself gives double the spell casting benefit of regular robes
of power, can be enchanted to +6/7, gives martial_bonus and an extra
attack and 50% chance of a second extra attack.