Fix: various issues with the Sword of Annihilation's attack. (EvilHack)
Couple things - monsters being killed by full on disintegration was
being handled incorrectly. Their inventory was being left behind, but
this is now properly handled. Also noticed that if a monster attacked
another monster with the sword, and a piece of armor was disintegrated,
sometimes the players armor was targeted. That's addressed too.

Note: In HackEM we don't have the SoA, but we do have the special disintegration attack available so this only updates that section.