The Iron Spoon of Liberation grants phasing with no timeout. Fixes #358.
This artifact has a bit of a different invoke from other artifacts. Most artifact invokes work with the property directly, but the IPoL uses a invoke that is not the same as the property.

For ex:
The Sceptre of Might has an invoke ability that is CONFLICT. This matches the CONFLICT property in prop_types. Same for The Heart of Ahriman, which uses LEVITATION and the Gauntlets of Defense, which uses INVIS.

The Iron Spoon (or Ball in other variants), uses a PHASING invoke to turn PASSES_WALLS on. But since it is not permanent , the invoke uses a timer instead to make it temporary (phasing is pretty powerful so this makes sense).

I'm going to try getting rid of the PHASING workaround and let the spoon work directly with PASSES_WALLS. This means that the invoke will not have no timeout!  Maybe this is overpowered, but we'll see. It means that invoking with phasing turns it off as it should, but also means that if the player drops the spoon, phasing is turned off. I debated making it mandatory to wield the spoon to get the phasing ability, but this feels like over-complicating things.