Revamp add_object_info for object lookups.
The object lookup code has been a bit of a mess. There were a few leaks in information with certain lookups (tins, armor appearances, etc), there was some inconsistency in how the headlines were displayed, and also dependence on the database entries for certain objects. This was an attempt to clean it up a bit and establish some order. I was able to fix the behavior for lookups when you don't possess the item so that you see all the info on an item type, but no appearance data leaked. The corpse and tin lookups also had a bunch of small bugs so fixed those, and the headline is correct now.

More work remains to be done, but this gets the core features functioning as they should again.