Players get the Shield Block technique when they reach skilled in Shields.
This is mostly to parallel how the disarm technique works. I moved some things around so that instead of roles just starting with shield block, they have to earn it by getting shield skill at least to skilled. This should be pretty easy for most roles, and it's a nice reward for training that skill. If shield skill goes back to basic or less due to amnesia, you also lose the tech.