Removed the Eye of Vecna and reverted the Hand of Vecna to SLASH'EM's behavior.
This is part of pushing more in the direction of old-school slashem. The Eye of Vecna is replaced by the Eye of the Beholder. The Hand has had all mechanics related to wearing it removed. I just prefer the original behavior more since I'm going for the slashem vibe. A couple things are convenient about this too, you don't need to wear the hand to get slotless cold resistance and 1/2 physical damage. I kept a few little things from EvilHack though - both can't be destroyed by lava, and the Eye uses the same death magic invoke that evil improved on. The Hand has the Summon Undead invoke from slashem, which needs a little bit of upgrading but I plan on keeping mostly the same.