Fix: Swimming under invisible ghosts caused Mon is not there panics. Fixes #322.
Really interesting bug. EvilHack commit 442a2ac22 introduced the concept of swimming under certain monsters in the water when trying to attack them. This bug arose because when in the water and trying to attack an invisible ghost, the player would get the 'You swim under' message, and an attack message. The problem is, we didn't actually swim under the monster, but the code seems to have still executed, resulting in us staying in our same position but the monster going somewhere else (presumably our position).

I added another check before swimming under so we don't switch places with invisible monsters. This bug didn't trigger when you could see invisible or even have warning so hopefully that suffices for now.