Remove potential trap creation when bomb explodes in water.
This was a strange find while I was trying to fix a different issue related to bombs. I noticed that the code from slashem has a path that can try to create a rust trap when the player is water walking above a water square when a bomb blows up. Removing this because it's not documented anywhere and has great potential for bugs.