Fix: Spellcasters can't cast Make Pool on squares with traps.
Mobileuser mentioned that he saw a pool created on a magic trap. This patch is attempt to fix that since it's weird if a trap and another feature can coexist on the same square.

I found a very similar function to the one I was going to write called zombie_can_dig, it just sees if the square is a room, corridor, or grave, and if it doesn't have traps. I just added grass to that list, and that is the list of eligible tile types that pools can be made on. I think this levels out the spell a bit more, plus it adds some benefits to things like scrolls of ice, which can be used to ice up tiles around you (and prevent pools).