Acquiring two alignment keys is now mandatory to ascend Vlad's Tower.
This took a bit of planning but I think this will work. Instead of using artifact doors (as originally done in SLASH'EM), I opted for the access blocking that EvilHack used for the Goblin Town. The artifact doors implementation is quite sloppy and would waste a lot of memory because every door would have to track it's status as an artifact door. In this implementation, you are barred from going up the stairs in Vlad's Tower until you get 2 alignment keys in your possession. Once you go up, you unlock that event and then the keys are no longer required. Access is also prevented by using levelport. I'm not sure of branchport yet, but we will get to that.

At first I was going to track the killing of the quest nemeses, but using the keys allows a player to steal them and retain their pacifist conduct if they want.