Two alignment keys can be forged together to create The Key of Access.
This is a fun use of artifact forging that gives the alignment keys a bit more spice. Any two can be combined. Also very important - the Key of Access can be used in place of two keys to get into Vlad's Tower.

I had to make a few changes to the alignment keys by converting them all to gemstone, this allows forging them. I also had the alignment keys on the totally indestructible list but this prevents them from being used in forging. So now, they can be susceptible to disintegration, but a very low probability. The Key of Access is indestructible however.