Upgrading/Tinker revisions.
The upgrade code needed a bit of love - the recent tweak that prevented flint rocks from being mass upgraded created some minor problems and not-accurate messages.
I improved the infrastructure so that we can get both the transformed object and the status of the transformation when upgrade_obj is used. Fixed behavior when inventory is full and the transformed object is dropped. Now only one object can be upgraded at a time, this applies to all objects not just flint rocks. We also now see the correct item and corresponding inventory letter when a stack is split.

I also noticed magical bag explosions had not been updated and would still delete contents, so now they scatter and deal some magic damage as well.