Fix: some 'gnomes hate eggs' tweaks.(from EvilHack)
Redo how a gnome will respond to seeing eggs in open inventory, so they
only check your inventory on the chance they do respond, not every
single turn. Only respond to regular eggs in open inventory, not all
types of eggs. Remove the 'screams/shrieks in terror' pline when a gnome
is hit by an egg; monflee() already handles that. Tweak the criteria for
when a gnome will respond to seeing eggs in open inventory (they can't
be asleep/confused/stunned/already fleeing/frozen).

Gnomish shopkeepers aren't affected by this, and gnomish priests don't
exist, so that concern is moot.