Various tweaks and cleanup for the wand of deluge.
While writing the wiki article for this wand I found a lot of loose ends. There was leftover code for the beam implementation, so I removed that and made sure all the ray code was called correctly. Also mirrored the behavior between monster and player interactions. The water also wasn't hitting objects on the ground (ie: scrolls and potions) so now it can be used as a convenient method of blanking items as was originally intended in SpliceHack. Steel golems are also soundly destroyed by this wand now - for some reason they take half damage, so even pumping 500 into them from a hit wasn't enough to crack their starting 290 HP. The muse code now correctly uses the ray instead of beam.