Scrolls of transmogrify can make materials the hero hates.
This restriction was in SpliceHack, but it's possible a player might want something they hate to use against other monsters. For example, an orc might want a mithril sword, a vampiric might want a silver dagger, etc. If the hero has gloves these items are usually not a problem. This also makes the results of transmogrification a little more unpredictable, which is more fun?