Fix: Flying stuff (From EvilHack)
Written by terrapin. In the course of testing the retouch_objects code, a bug was found in how taking off armor was handled when you had wings. This had a simple one-line fix, but in the course of testing it, another bug was found in how armor was put *on* when you were winged, and more generally how hard body armor and big wings interacted with other kinds of flying. Fixing this involved tinkering with the definition of Flying, which I'm not a huge fan of, but I tested different sources of flying and BFlying extensively after the changes, so here's hoping it holds up.

hackemslashem: I also saw that some code was duplicated for the cloak of flight and celestial dragon scales so I was able to consolidate that.