Allow tipping into containers on the ground (reported by hothraxxa). Fixes #433.
My fix was basically to copy the whole block for inventory tipping and use it to check on the ground. I figured out how to iterate over floor items by looking at the code for burn_floor_objects(). Could use a refactor in the future but it seems to be working now.