Bombs fuses are snuffed if monsters go through portals or migrate levels.
This is to prevent awkwardness after a monster migrates levels with an armed bomb. The bomb's timer won't migrate correctly, so they just end up with a permanently armed bomb that never blows up. There might be alternative solutions like making the bomb blow but that can result in other bombs igniting, so this seems the simplest.