Revert: Scrolls of charging and bag of tricks.
This fun and innocent looking patch from UnNetHack unfortunately has some complex implications and it's better off just left out. If a chain of items are put in the bag, with the scroll of charging first, we can end up with a charged and non-empty BoT. There's also the issue with tipping a charging scroll into a BoT.

I also added a sanity check for bags of tricks and rats to prevent charged and non-empty bags from existing.