The Wizard of Yendor will never steal the quest artifact of the current role (from UnNetHack).
This is something I have wanted to implement for a while but just didn't have time or energy - so I was delighted when bhaak was able to pull it off - mega thanks to UnNetHack on this one!

I think the premise of the hero's quest artifact being special and non-stealable is strong - there should be a certain unbreakable bond between the two. It's my hope that this change makes many role quests more enjoyable since you'll actually *use* your QA in the endgame and be able to rely on it when trying for more conducts, etc.