Handle life-saving when a monster offers the Amulet
If your life is saved when a monster sacrifices the Amulet of Yendor,
demons and the Wizard kill you again, while others allow you to live. In
any case, the game ends. Also, a coaligned or generally friendly demigod
may not use their power to kill you, same as if you're extremely lucky.