Wearing a helm of telepathy can focus psychic attacks, tweak to ESP object property (From EvilHack).
A player as an Illithid wearing a helm of telepathy made of the right
material has a chance of the helm focusing their psychic attack and
causing more damage than normally. I mention 'made of the right
material' because helms made of any metal except mithril (Mitre of
Holiness is an exception) block an Illithids ability to use its psychic
attack. The extra damage is 3-8 hit points.

Also in this commit - I've seen it a few times now, a player wishes for
a piece of head gear with the ESP object property, and specify
'telepathy' instead, except a helm of telepathy already exists, and it's
caused some issues and confusion. Current solution - ESP object property
is disallowed on head gear, helm of telepathy already exists and can
spawn made of a material that doesn't interfere with spellcasting or
psychic abilities.