Adjust some dragon-scaled secondary resistances/passive defenses (From EvilHack).
hackem: This patch was roughly adapted from EvilHack. While doing so, I realized that we are getting fairly close in how the dragon scales are acting so I decided to adapt most of the changes in an attempt to unify the behavior between EvilHack and Hack'EM dragon scales.

* Yellow dragon-scaled armor secondary effect - We had this already but the evilhack implementation was a bit more thorough.
* Orange dragon-scaled armor - EvilHack matched hackem.
* Shimmering dragon-scaled armor removes stealth as a secondary effect
and adds stun resistance (this also makes shimmering dragons stun
resistant)
* Shimmering dragons breath attack - I did NOT add this. We are already maxed out on the zap types and I did not have time or resources to add this. The shimmering dragon already has a passive stun so that will have to suffice.