Fix: gaining certain resistances would not cure current effect they protect against / Snickersnee grants stun resistance when wielded (From EvilHack).
This became a 'twofer' commit. Was in the process of adding stun
resistance when wielding Snickersnee (nice idea krm26) when I noticed
that having resistances to certain conditions would not fix/cure those
conditions if the player was already affected by them. Some code was
already in place for curing stoning if donning yellow/chromatic-scaled
armor, but not for being stunned or sick. this commit addresses that.

Also noticed a couple places where 0 should have been expressed as 0L.