Removed sickness resistance from Plague.
Slotless sickness resistance is just too good - in dNetHack there might be a reason for it, but it bypasses some huge endgame challenges. For now I went back to the standard wielding-grants-poison resistance, but in the future I think we could make it more true to it's name and convert it to shoot diseased arrows (similar to Grimtooth) and instead grant sickness res when wielded.