Renamed boxing gloves, fixed stun mechanic. Fixes #473.
This fixes a weird situation with the 'boiled gloves' description. So instead, the boxing gloves are now known as the gauntlets of force, and the boiled appearance has been replaced with the boxing gloves. I also noticed a bug with the stunning mechanic when wearing the gloves (it would trigger 24/25 times instead of 1 in 25!), so I improved it a bit so the odds are 1 in 20 now.