Circe's Witchstaff buffs. I wanted to make this item a bit more interesting and not just an automatic throwaway for min-maxers. Basically I'm playing with the idea of rewarding some useless genocides. When pigs and feral hogs are genocided, the special ability will instakill it's target instead of polymorphing them into a pig. I'm picturing what would happen if the pig transformation were forced, but since no options exist they are contorted beyond death.
I bumped up the chance of the special from 10% to 20%, and doubled the to-hit bonus. I also removed the taming invoke ability since it just duplicates what the taming scrolls can do.

The major weakness of this weapon is it isn't effective against anything with magic resistance. Hopefully being a Staff of Matter helps balance that out.