Cleanup makekops and imported kop kommissioners (from SlashTHEM).
While investigating issue #479, I discovered some of the kop code wasn't quite right. I had used some of the code from Unnethack but was able to adjust it back to the better 3.6. I also brought in the kommissioners because we had generals (for the black market summons) but no 5th level kop.