Reduced the size of the Black Market and brought it back.
A big problem with the black market in SLASH'EM is... it's just too big. I had already narrowed it slightly when I brought it into Hack'EM, but it creates two problems: 1) It gives the player too large of a item dump and a huge advantage regarding resources (especially if they are able to figure out how to steal) and 2) it creates a lot of tedium in navigating all the loot.
I also removed the UnNetHack version.