Removed forged weapons.
Removed Squall (it was already too similar to Thunderstruck). Gave Thunderstruck double damage, and it sends a lightning bolt instead of sonic ray.

Mortality Dial will remain but I reduced the tohit and dmg bonus by half.

The only artifact which can currently be forged is the Key of Access.