Vampiric hunger and regeneration tweaks.
Okay, this is another mechanic that's been through the idea bin many times, but I think we are finally converging on a good setup. Instead of the HP gain while feeding thing, we are ditching that and going with good 'ol Regeneration from XP 1. However, we are also introducing Hunger at XP 1. Similar to the giant race, vampires do not gain extra hunger from regeneration, so I think this is necessary to give the player an incentive to play fast and aggressively, truly utilizing their regeneration to its full effect.

Vampires are also not punished for being satiated all the time, so you can have a feeding frenzy and hopefully make it last for a while, allowing you to search for more victims.

I moved the acquisition of sleep resistance up to XP 9. This is a bit unusual because the only variant generous enough to grant it is dNetHack. I believe it is essential though, given that vampires cannot eat for intrinsics, but it shouldn't be available at XP 1. Vampires already have Unbreathing, allowing them to bypass sleep traps with ease. By the time you are close to XP 9, you will truly need it to deal with threats like kamadan and random wands of sleep.