The Hand of Vecna weighs 200.
It's a heavy burden to bear, a true source of evil that drags most adventurers down with it. The hand is quite powerful, it it's powers should have some weight to them - literally. Specifically, 50 for half-physical damage, 50 for slotless cold resistance, 50 for hungerless regeneration, 25 for the invoke, and 25 for the ability to wield it and gain drain resistance. The Hand is guaranteed in every game of Hack'EM, so I challenge the player to some inventory management when confronted with Vecna's hand.