New upgrade paths. Although I like the original setup in SLASH'EM, we need to do some tweaking for a couple reasons: 1) I want the object lookup to be able to show an objects upgrade info easily, and this will be much easier if each item only has a 1-to-1 relationship (mapping) to the item it upgrades into. 2) Since properties transfer into their new upgrades (much like forging), we *have* to hold back a bit on the magical items that can be upgraded. I believe retaining bags of holding is a must, it relieves a lot of frustration for players, but things like speed boots, cloaks of magic resistance, etc - I'm going to drop. As a positive, any items that are considered junky I've tried to give a decent magical alternative if one exists (ie: the helm of madness changes into a helm of brilliance).
I originally dropped a lot of the weapon upgrades because they were a bit redundant by having alternatives in the forging recipes, but I've re-added those and used the most obvious patterns where possible.

The 6 magical staves present a special problem, so my solution is to pair them up with
each other. The regular quarterstaff can be upgraded into a silver capped staff, but not
into a magical staff. A similar pattern is being applied to robes, so a regular robe can
only be tinkered with a large robe and vice versa. The magical robes have a circular
upgrade pattern (weakness -> power -> protection -> weakness...)

The gray stone upgrades were also a unique situation. I'm going to try and let loadstones be upgraded into healthstones because it adds value to loadstones, but also makes the process of conversion a bit more challenging. I removed flint from the path because it's just too common. Loadstones are also not uncommon, there are usually some to be found in mine's end (especially the Boulder Bonanza level), so this might make mine diving even more rewarding.