Removed the ring of psychic resistance.
This was a bit of an unfinished import from SpliceHack. It was meant to deal with more psychic related threats (gnomish psychers and some other AD_PSYC attacks) but these didn't really port in well from splice). Right now the only thing this item is good for is easy removal of the afraid status - which isn't really enough for me to keep it. Players are encouraged to drink booze to combat fear.