Wielding banes will only anger monsters in view; shopkeepers will ban you. Before wielding a bane would render all the hated monsters on the level hostile no matter their location.
We also use setmangry, which gives the player an indication something has happened. It will also ding their alignment for angering peacefuls.

Shopkeepers will not become angry, but will ban you for life if your bane matches their race. This is similar to how convicts are banned and uses the same shopkeeper flag. An interesting note is that this fits perfectly with existing store hawker code, so instead of inviting you in, they openly show their contempt for you.