Tweaks to monster behavior when wielding/wearing a bane artifact.
From EvilHack commit f9a7db8a185463d8323455ae7bcb873845bf19a7.

This builds on previous commit 30785e44d. In that commit I neglected to give credit to K2 for idea for shopkeepers banning the player. Credit also to hothraxxa for reporting the behavior of banes alerting all monsters on the level regardless of sight to me in the first place.