Scroll of magic detection nerf.
When blessed, this scroll would act as a blanket type identify for all magical items on the level. Scrolls of identify and shopkeeper identify services pale in comparison to this effect, so I've limited the detection to just the magic and properties. A blessed scroll will still reveal if all items on the level are magical so it's definitely still useful for sorting through masses of items.