Adjusted odds of removing properties. When paying a shopkeeper to do it, there is a 1 in 4 chance of success (this is how premier identify worked in SLASH'EM). When dipping into a potion of amnesia, there is the same chance as being hit by a thrown potion (1 in 13).
This is to solidify the scroll of transmogrify as the most reliable means of property removal - which will be a fairly important procedure to perform with the new system of properties. Amnesia is trivial to create bulk quantities of and shopkeeper services are an unlimited resource (if you are lucky enough to have one in your game) - hence the slight adjustments.