Overhauled the fear/afraid system; new fearless property for weapons and armor.
I've been a bit unhappy with how the fear stuff has been handled, it was very clumsy and inconsistent. So I cleaned it up and gave it an official attribute called Fearless. This attribute is now de-coupled from the Psychic_resistance attribute. Before, illithids would start the game with both psychic and full fear resistance!

I gave the necromancer intrinsic fear resistance from level 1 - feels like the right role for it. I also removed the undead slayer from any fearless parts, they will have to find their own sources of bravery.

Pridwin grants full fearlessness when worn and Dragonbane will protect you from - and cure - the frightful roars of Dragons, but only dragons. I decided to leave the tinfoil hat as is, so it still grants psychic and fearless - but it blocks ESP and clairvoyance, a tradeoff the player will need to contend with.

I also added another party trick to the magic trap that can cause a very short period of fright.

The new bravery property should also come in handy. Weapons and armor with this property will prevent fear, and can be worn or wielded to cure the afraid status.