If the player makes 13+ wishes - luck is locked at -13. I tried experimenting with letting the base luck be -13, but it was very difficult to keep consistent between saves because there are lots of different things that can affect luck between games (friday13, new moon, etc). So this turns out to be the most reliable and bugfree method to execute this idea. Even though the possibility of getting 13 wishes is unlikely, I already think this might be the start of a new conduct...
Since this commit may attract attention, I'd like to explain the rationale behind it. I don't have a strong aversion to wishes, but I also don't necessarily favor them. They are a staple of NetHack; very few other games offer a similar feature. However, wishes are undeniably overpowered. Recognizing that many players enjoy wishes as a fundamental aspect of NetHack, I've chosen not to limit the number of wishes or the types of wishes players can make. Instead, I'm introducing a gradual and subtle effect as players make more wishes. The first wish remains unaffected, but from the second wish onwards, the game becomes progressively more challenging. My goal is to introduce compelling decisions for the mid-to-late game and push players slightly out of their comfort zone. I anticipate that most players will still make 3-4 wishes, roughly cancelling out the effect from a luckstone.

I selected prime numbers as the thresholds for the luck penalties because they are inherently fascinating, and they seemed to align perfectly with the intended progression.

This will be the last wishing nerf - I'm interested to see how it pans out. My ultimate hope is that it creates more interesting stories and talking points for both deaths and ascensions. Maybe someone was able to save all their wishes for the astral plane, and then cashed them in to avoid trouble on the ascension run, or maybe somebody deliberately farmed 13 wishes to ascend with -13 luck, or maybe someone tried wishless for the first time to avoid any trouble at all! I don't know what the results will be, but when people play Hack'EM I want to keep them thinking and on their toes!