Adjustments to Blade (Und-Vam)
I rejiggered some of the handling of the Und-Vam inventory. Instead of denying some of the food items, I just substituded blood for holy wafers. It's possible the player can use garlic to ward off undead (Blade had holy garlic water in the movies), or use the wolfsbane to tame or pacify monsters. I also made the trench coat a bit weaker, it doesn't grant any AC and gets no enchantment bonus - but it does grant fire resistance. This is to counteract the vampires innate vulnerability to fire. It also makes the trench coat stand out a bit more and could be useful for other roles.

I was also able to make the role start out with their lenses pre-worn, which makes for one less chore at the beginning of the game.

The vampire starts out with innate regeneration, so I think it's fine to make this particular role/race combo be more offense heavy and reduce some of their defense in compensation. The steel long sword is also much more powerful than the wooden stake the other undead slayers start with.