Fix: Clashes with acid/sonic resistance and vulnerability to those resistances.
I found it was possible to have acid resistance and vulnerability to acid, but not take any acid damage at all. I think this goes against the purpose of the vulnerability mechanic so I updated all the Acid/Sonic resistance checks with how_resistant - using 50 as the baseline. These resistances are not partial, so we can only really use 50 anyway. If a player has acid resistance and is vulnerable, their "effective" resistance is 50.

This commit may require a followup to make sure any updates have corresponding resist_reduce usage to prevent full blast where perhaps 50% is warranted.