Refactor rndmonst(). From EvilHack commit 76eacd0b14.
The initial intent of this commit was to *gasp* decrease how often
zombies spawn. Because there are so many zombie templates for rndmonst()
to land on, the odds are in favor of some type of zombie spawning vs
something else. Was discussing this with qt, and he suggested that I
refactor rndmonst() the way it's done in vanilla NetHack 3.7 (which was
an aosdict/xNetHack commit). So this commit does that, which allows
player monster handling to be expressed in a better fashion. And then I
noticed that most of the lower level zombies had a spawn frequency of 2
(monst.c), so instead of having to make a special case for them in
rndmonst(), it's better to just drop that spawn frequency from 2 to 1.

Long story short, zombie spawn frequency is reduced.

With this refactor, I took an opportunity to add a special thing for
Halloween, which I think was mentioned in IRC a long time ago, mostly as
a joke. On Halloween, the odds of any undead spawning increases
threefold. The player will also receive a notice of this every time they
load their game on that day, just like when there's a full/new moon.

---
Minor change for hackem: I left the group size of the zombies alone.