The tumble technique can confuse monsters after tumbling past them. The chance depends on if they are innate flankers, since outflankers would be better at anticipating your actions. Displacing a monster into a trap always confuses them.
I also noticed that the handling for monsters was lacking checks for traps or displacing them into water (if possible). I was able to check for both of these conditions, however, if monsters are killed this way they currently do not give the player experience for doing so. This will require some extensive refactoring to use the existing displacement code and is a project for a different time.