Drain life refactoring and fixes. After fixing up some of the hydra mechanics - I saw that we needed a centralized method for draining a monster's life (similar to how the player has losexp). Each time a hydra gets drained, it should lose a head - this commit addresses that and also consolidates a lot of the draining code for monsters.
I also discovered a bug where the damage for a wand of draining and a couple of other drain life attacks were having damage arbitrarily doubled. The vampiric bite followed the same pattern, but I kept the damage addition because without it they wouldn't have as potent (or fun) of an attack.