Created a droid pet for the jedi role. This was really fun to make - it's my hope that this makes the jedi role even better and solidifies it as the 'newb barb' role for new Hack'EM players.
The droid is quite powerful, albeit slow, but will prove a very useful companion to have around. It has resistances to almost everything, regenerates, and flies, so it should be very durable. The droid does not leave a corpse, so onces it's gone - it's gone, you will want to protect it from destruction! Keep in mind its main vulnerability is shock damage. If it runs into a web - it has a tiny saw that it can cut away web traps with.

For offensive capabilities, it has a strong melee zap touch and a ranged shocking ray zap.

However, one of the main features of the droid is it's ability to automatically BUC test and property test any items it walks over. Since it's a robot - it also never hungers, and it's tameness will never expire. So you can leave it in Sokobon forever as a item processing tool, or you can level it up and bring it out into the wild. Note that since it doesn't have hands it cannot pick up items like other pets - it probably would not be very good at stealing from shops anyway since it's so slow. A good first goal for the jedi will be finding a wand of speed monster to bump up its speed.

This droid will need a bit of playtesting, but I think we have a decent start. Other ideas that could be fun are adding: oil slick spray to blind enemies, fire extinguisher, or defensive passive saw attacks, but let's start simple and build from there.