Fix: Monsters could drain life fern spores and continue their attack chain. Fixes issues #479 and #19.
Issue 479 was: dmonsfree: 36 removed doesn't match 37 pending on Dlvl:13. What I found specifically with the fuzzer was that a slaughter wight was attacking a fern spore, draining it to death, it exploded (due to spore death), and then the slaughter wight continued on with the rest of its attacks. The death by draining wasn't handling the death correctly so two deaths were reported. This also revealed how the ancient issue #19 was happening, a light-source-emitting monster was simply dying in the same fashion!