Revert dce86f3f2 - Quest leaders will make you forfeit the quest artifact only if you heavily abuse alignment (less than -64 alignment abuse score required)
Returned to the EvilHack style of quest alignment abuse handling. I think it just works a bit better that way, giving the player until -64 abuse practically guarantees they won't have to fight their quest leader. I relaxed the requirements for the boon to be better than -13 alignment abuse. If you get the boon, your current luck is also maxed out.

However, if you trigger a fight with the leader, you are out of the running for the boon.