Enabled the sleep and sickness properties as offensive properties on weapons. This makes these properties conform the existing pattern for resistance properties: weapons have the offensive effect, and armor/rings confer the resistance.
On weapons, the sleep prop activates 1 in 3 times. This is so the Drowsing Rod is still the king of sleep. Otherwise, it works similar to the Drowsing Rod effects. It also doesn't get the usual bonus damage other properties do.

The sickness property was renamed to filth for weapons, and it works similar to Grimtooth. It does receive a small amount of bonus damage, but only 1d2 on hits.
When armor has the sickness resistance property, it's known as the health property.