Overhaul of ring appearances and materials. This fixes issue #504 - rings could not have their material manipulated by the scroll of transmogrify. This required a complete reset of the ring appearances and the appearances have been decoupled from the materials. The main problem was that many of the ring appearances *were* materials, and changing the actual material would result in paradoxial situations (ex: you might try to change a silver ring to a gold ring, but a gold ring already exists.
I also updated the display of rings so their material always shows in front of the appearance or ring type, and the scroll of transmogrify can be used to change a ring to a random material now.

Thanks to hothraxxa for mentioning how the previous behavior could result in a vampiric player getting the unfortunate silver ring of slow digestion - at least now the material can be changed!