Fix: Summoned spheres stay indefinitely if abandoned From EvilHack commit a3efd3a71c4
Pets from the freeze/flame sphere spells only stay around for 100 turns or so, kept track of via msummoned. When it hits 1, the monster vanishes. When returning to a level that had summoned spheres, if you stayed away for longer than they were supposed to survive, then msummoned is set to 1. But, it is decremented in monmove.c, making msummoned 0, before the check for whether it's timed out is made. This makes abandoned spheres become normal pets. Instead, set msummoned to 2, so they wink out of existence properly.

It is a bit weird that the spheres wink out of existence immediately when you return. But, neither mon_catchup_elapsed_time nor its callers are equipped to handle getting rid of monsters, so it would be difficult to make the monsters disappear silently without some significant changes. Another option would be to treat it like withering and totally ignore it---it is still odd, but less obviously so than getting spammed with messages as soon as you arrive on the level. My impression is that no one uses these spells enough to notice either way though (their loss!).