Altar nerfs, modified from SpliceHack.
After two artifact gifts have been bestowed, there is a chance an altar will be destroyed on every subsequent gift. The chance does not start as high as SpliceHack did, but it will grow with the number of gifts. Minion pets do count as gifts, but they will not have a chance to immediately destroy the altar.

This is part of my agenda to push and prod player further into the dungeon, whether they like it or not. Splicehack had tried (and reverted) this change, but I believe it will stand as a hardcore fundamental of Hack'EM. We have at least 3 guaranteed altars in the main dungeon: Minetown, the Lawful Quest, and Moloch's Temple. Many roles also have altars. The altar should be viewed as a limited game resource - not an infinite resource. I believe that players should play 'the hand that they are dealt'.

Side note regarding Infidels. Infidels pose a particular problem since they rely on altars as part of their core gameplay. Therefore, Infidels will not suffer breakage of altars, but neither will they benefit from altar gifts beyond getting Secespita.