Jedi and Undead Slayers don't train martial arts, they train bare handed combat.
This change is to prevent some role dilution I felt creeping in. The masters of martial arts are monks and samurai, and I think letting Jedi in particular gives them a bit of an 'out' of their narrow selection of weapon skills. The undead slayer already has had their share of buffs, and I don't think this one particularly fits anymore.