When learning weapons through use (becoming familiar with them), you fully identify the item and also discover the object's base type.
I noticed in playtesting that the player could become familiar with a weapon they didn't start the game with knowledge of (ie: a caveman doesn't start with knowledge of sabers), but then become familiar with a saber and still have the pokedex not show the base damage versus monsters. I also decided to grant a full identify because this does take time and it makes sense that a player would know the ins and outs (BUC, properties, all of it) after gaining such familiarity.