Weapon cross-training (from DynaHack) From DynaHack commit ff0ed9b66a.
High-level weapon skills will accelerate the training of lesser skills
in the same skill groups by x2/x3/x4 (shown in the #enhance menu),
depending on the skill level difference.

The weapon skill groups are:

 * Short Blades: dagger, knife, short sword
 * Chopping Blades: axe, pick-axe, broadsword
 * Swords: short sword, broadsword, long sword, two-handed sword,
   scimitar, saber
 * Bludgeons: club, mace, morning star, flail, hammer, quarterstaff
 * Flails: morning star, flail, whip
 * Polearms: quarterstaff, polearm, spear, javelin, trident, lance,
   unicorn horn
 * Launchers: bow, sling, crossbow
 * Thrown: javelin, sling, dart, shuriken, boomerang

The skills that belong to more than one skill group are:

 * short sword: Short Blades, Swords
 * broadsword: Chopping Blades, Swords
 * morning star: Bludgeons, Flails
 * flail: Bludgeons, Flails
 * quarterstaff: Bludgeons, Polearms
 * javelin: Polearms, Thrown
 * sling: Launchers, Thrown

This counteracts the large training point increase of UnNetHack, and
means that advancing weapon skills that don't have the highest damage
output for the late-game is no longer a complete waste of time.

This commit also removes skill menu code that is unused in NitroHack.

See also Karadoc's K-Mod NetHack variant for a different take on skill
cross-training.