Magic chests (from EvilHack).
Introducing the very popular and powerful magic chests from EvilHack. I had to think for a while on whether to import these. I originally had the position that creating stashes is a Very NetHacky thing to do... but after seeing the positive reaction to them I reconsidered. I also started to realize that with all the property manipulation that is possible - having a convenient place to consolidate gear would be an absolute godsend and make the early discovery of properties much easier - therefore making the game more fair and tipping the scales back in the player's favor. The implementation is virtually identical to EvilHack's except for where our content diverges.

I had to make a few adjustments: magic keys won't be found in Gollum's cave (doesn't exist in Hack'EM), but instead can be found in the Village level in the thug's hideout. I also moved the magic key at Fort Ludios to Grund's Stronghold. All the Mine's End variants that are not in EvilHack also got their own magic key placed somewhere sneaky. All the extra roles also needed magic chests placed in their quests. The magic chest at Rodney's tower is on the bottom level.

Since the dungeon is bigger, I added a couple more magic chests for the players convenience to: the black market, the temple of moloch, the pleasant valley village, and the lawful quest.

This is a squashed commit of all the following changes from EvilHack.

[x] f20b19867 - Fix: crash when trying to loot a magic chest with nothing in inventory.
[x] 08ffd005d - Fix: magic chest placement and quest leaders.
[x] 07ff80000 - Fix: don't allow magic keys in bones.
[x] 1f3a3e391 - Fix: Give message when walking onto a magic chest square.
[x] a674a3063 - Fix: Report objects on magic chest square when nearlooking
[x] f1fe3c6b1 - Fix: Occasional program in disorder when autounlocking magic chests
[x] 7a4121193 - Fix: farlooking tools
[-] 8b2e3b012 - Change placement of magic key in Gollum's Cave.
[x] 150f482d3 - Fix: don't prompt to autounlock a magic chest if you don't have a magic key.
[x] 234aa358d - Magic key placement.
[x] 5c4e24cad - Fix: prevent monsters from hovering over a container with something in it that they want but can't carry
[x] b7b0fcab1 - Autounlocking of magic chests
[x] 7b6b69554 - autokey() tweaks
[x] 35dd2dff1 - Tweaks and fixes to magic chests.
[x] 06359fe9f - Monsters can loot magic chests Monsters will path to and loot magic chests much like normal chests. If it is locked, they must be carrying a magic key in order to open it.
[x]29a7dd12d - Spells/wands zapped down at a magic chest will (un)lock it if applicable Whoops.
[x] 4aaa75463 - Partial fix: monsters can loot bags on the ground that they can't carry.
[x] 1464c4a98 - Magic chests can be locked and unlocked
[x] d6bbf4c91 - Magic chest placement.
[x] aa8e240e0 - Tiles IV: The Quest for 73s
[x] 7beaf7a2e - Tiles III: Revenge of the Switch
[x] 2fcc5f0af - Tile Redux
[x] 6d189a587 - New tile for hidden chest Are you happy now, Windows?
[x] bfa17cf54 - Finish magic chests
[x] 98e8e4318 - Flesh out magic chest interactions
[x] 3b0e7bfed - New object: "hidden chest"
[x] 6cba3ee32 - New dungeon feature: magic chests