Allow all position prompts to be aborted (DynaHack).
This applies to: jumping, polearm pounding, grappling hooks, scrolls of
stinking cloud, paying with multiple shopkeepers in sight and spells of
fireball and cone of cold.

All of these code paths had the ability to handle the position choice
being aborted by the player, but the `force` argument to `getpos()` was
preventing that from ever happening.