Fix: Adding properties with transmogrification or shopkeeper service. There were a couple of problems with the first version of adding properties. If the item was magical, it would only get one 10% shot at adding a property, which is quite a letdown for an expensive service or writing a scroll solely for that purpose. Now when adding properties, it will loop 1000 times until it has success adding a property, so that magical items are also viable.
I updated the shopkeeper service a bit - non-magical items are cheaper to add properties to ($850), and magical items are more expensive ($2250), but at least you know a property is almost guaranteed - even if it is a negative property. The service will also conveniently reveal the property so you get immediate feedback on your purchase.

Finally, I updated the scroll of transmogrify so that only adding properties to magical items requires the luck check. This follows the shopkeeper pricing, and lets players use the scroll more conveniently with non-magical items.